+++ src/java/servlets/administracion/AgregarProductoServlet.java
@@ -108,9 +108,9 @@ public class AgregarProductoServlet extends HttpServlet {
         
         //prod.setProductID();
         prod.setProductName(nombre);
-        prod.setProductPrice(Float.parseFloat(precio));
+        prod.setProductPrice(Integer.parseInt(precio));
         prod.setProductLongDesc(descripcion);
-        prod.setProductStock(Float.parseFloat(stock));
+        prod.setProductStock(Integer.parseInt(stock));
         prod.setProductImage(imagen);
         
         ModeloProducts productoModel = new ModeloProducts();
